Introducing owned suqres computation.
Uses attackedBy and attackedBy2 for all piece types. Outputs ownMove and
ownDef bitboards.

For pawn moves (since they do not attack the square they move on), check
for owning the move as pushSquare & ownDef. Assuming our pawn is not
blocked, we own a pawn move if we can defend the pawn afterwards.

ownMove excludes all squares covered by opponents pawns and so is
suitable for piece moves only.