Add doubly pawn-attacked squares to fix pawns
If a doubly pawn-attacked square is a candidate to become untouchable
(and thus potentially fixing pawns), this inherently means that it is
not in attack span of the opponents pawns or or all pawns able to attack
it are themselves already fixed. Thus, breaking thus pseudo-fix requires
the opponent to sacrifice the pawn entirely. This in turn means that
that pawn's attack span is indeed non-existent.