Use averages in closed positions
The psq values seem to be involuntarily biased quite close to zeroing
expectation in this closedpos bench. I speculate that these positions
are more naturally aligned with what matters for playing strength and
that the long tuning process for the existing values was biased this way
to drive expectation to 0 naturally.
This commit enforces this behaviour deliberately.

Bench: 3767487