Change use of 3rd and 4th value for winning adj
Change mergeClosedness() to leadershipAdjustment().
leadershipAdjustment() adds a second real \in [-1, 1] dimension of
winning to the MG+EG value concept.

Bench at this point is still equivalent to master. This is expected
since all 3rd and 4th values are still zero.

THis makes use of the QScore implementation using the terms "closed
game" and "open game" of the additional QScore values. These terms no
longer reflect their usage within this test.