Add tunable parametrization for eval params
Preparing some non-Score-type parameters in evaluate.cpp and pawns.cpp
for QScore tuning.

The new tunable parameters in pawns.cpp:
  - "Shelter"
  - "Support"
The new tunable parameters in evaluate.cpp:
  - "estimateNorm" (more dangerous to tune)
  - "Kd"
  - "Pass"
  - "PassFree"
  - "Space"

The provided tuning intervals are such that at the boundaries, the
original value is negated (-128) or doubled (+128).
EstimateNorm tuning should be very conservative with bounds.

Bench: 4923208